Fix pre-commit hooks for helm and shellcheck
This change fixes the pre-commit configuration to correctly run `helm lint` and `shellcheck`.

The `helmlint` hook from `gruntwork-io/pre-commit` is deprecated and was causing pre-commit checks to fail. This has been replaced with a local hook that runs `helm lint` on the chart directory.

The `shellcheck` hook was also failing because the executable was not in the `PATH`. This has been fixed by ensuring the pre-commit environment has access to the `shellcheck` executable.