feat(graphql): Allow custom selection sets
This change introduces the ability for users to specify a custom `selectionSet` for GraphQL queries.

Previously, the GraphQL upstream would always generate a selection set that included all possible fields. This was inefficient and defeated one of the main purposes of GraphQL.

Now, users can specify a custom `selectionSet` in the service configuration. If a `selectionSet` is provided, it will be used to build the GraphQL query. If no `selectionSet` is provided, the old behavior of including all fields is retained.

This change also includes:

- A new test case to verify that the custom selection set is used when specified.
- An update to the documentation to explain how to use the new feature.