feat(config): Add support for environment variable secrets
This change adds support for sourcing secrets from environment variables.

Previously, secrets could only be specified as plaintext values in the configuration file. This is not ideal for production environments, as it exposes secrets in plain text.

This change introduces a new `environmentVariable` field to the `SecretValue` message. When this field is set, the server will read the value of the specified environment variable and use it as the secret.

If the specified environment variable is not set, the server will fail to start. This prevents the server from starting in an insecure state.

The documentation has been updated to explain how to use this new feature.